rework: Organize components
